+++ src/gpu/gpu_oei_assembler.h
@@ -11,9 +11,9 @@
    !_____________________________________________________________________!
    */
 
-__device__ static inline void OEint_vertical(int I, int J, 
+__device__ static inline void OEint_vertical(uint8_t I, uint8_t J, 
 #if defined(DEBUG_OEI)
-        int II, int JJ,
+        uint32_t II, uint32_t JJ,
 #endif
         QUICKDouble PAx, QUICKDouble PAy, QUICKDouble PAz,
         QUICKDouble PBx, QUICKDouble PBy, QUICKDouble PBz,
